requirements.txt
Django>=6.0.1
psycopg2-binary>=2.9.11
python-dotenv>=1.2
gunicorn>=24.0
django-unfold>=0.77
openpyxl>=3.1
django-mptt>=0.18
